Add files via upload
The objective of the final project is to implement K-means algorithm for the Wisconsin Breast Cancer data set using Python.

Breast cancer is a rising issue among women. A cancer’s stage is a crucial factor in deciding what treatment options to recommend, and in determining the patient’s prognosis. Today, in the United States, approximately one in eight women over their lifetime has a risk of developing breast cancer. An analysis of the most recent data has shown that the survival rate is 88% after 5 years of diagnosis and 80% after 10 years of diagnosis. With early detection and treatment, it is possible that this type of cancer will go into remission. In such a case, the worst fear of a cancer patient is the recurrence of the cancer.

In the final project, we will address this issue by using k-means clustering to analyze cancer data and classify patients into two different groups, one with benign and the other one with malignant cells. You will implement K-means clustering program in Python and test it on famous Wisconsin Breast Cancer Data.